provider/aws: Separate existing validate functions
