+++ builtin/providers/aws/resource_aws_dynamodb_table.go
@@ -801,18 +801,3 @@ func waitForTableToBeActive(tableName string, meta interface{}) error {
 	return nil
 
 }
-
-func validateStreamViewType(v interface{}, k string) (ws []string, errors []error) {
-	value := v.(string)
-	viewTypes := map[string]bool{
-		"KEYS_ONLY":          true,
-		"NEW_IMAGE":          true,
-		"OLD_IMAGE":          true,
-		"NEW_AND_OLD_IMAGES": true,
-	}
-
-	if !viewTypes[value] {
-		errors = append(errors, fmt.Errorf("%q be a valid DynamoDB StreamViewType", k))
-	}
-	return
-}
